requirements.txt
openai>=1.0.0
anthropic>=0.25.0
gitpython>=3.1.40
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
